requirements.txt
flask==3.0.3
groq==0.9.0
requests==2.32.3
python-dotenv==1.0.1